refactor(backend): move local mod translation logic to a new mod, resolve clippy warning
