+++ src-tauri/src/instance/helpers/mods/mod.rs
@@ -4,3 +4,4 @@ pub mod forge;
 pub mod legacy_forge;
 pub mod liteloader;
 pub mod quilt;
+pub mod translation;
